requirements.txt
certifi==2024.8.30
click==8.1.7
influxdb-client==1.48.0
python-dateutil==2.9.0.post0
reactivex==4.0.4
six==1.17.0
typing_extensions==4.12.2
urllib3==2.2.3
w1thermsensor==2.3.0
